Refactoring out the app creation
